hotfix for appending multibodies
when appending multibodies the wrong ndof was used which caused a matrix slicing error.
this is a hot fix which takes the correct ndof but it naively copies the val, damp and acc
and does not take the correct ndof row. Also the internal index was wrong for the child links in the appending multibody